Version 1 Main Menu
Finished the first prototype of the main menu for the game. This will be improved as I get the art I need. I also still need to add remaining credits that I have no received yet. Again, this is not final, but it works.

Note: In the Main Menu Container, you can set what the first level should be when the player presses the "Play" button ;)